pyproject.toml
[tool.poetry]
name = "timeseries-migrations"
version = "0.1.0"
description = ""
authors = ["Karl <karlwolffgang@googlemail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
gewv-timeseries-client = "^0.3.3"
"fs.smbfs" = "^1.0.1"
python-dotenv = "^0.17.1"
"fs.sshfs" = "^1.0.0"
influxdb = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
black = "^21.5b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
